refactor: separate navigation and generated page assets
