+++ scripts/docs-site/build.mjs
@@ -2,7 +2,6 @@
 import fs from "node:fs";
 import path from "node:path";
 import { createHash } from "node:crypto";
-import { Worker } from "node:worker_threads";
 import { parseArgs } from "node:util";
 
 import { ignoredDocDirs, ignoredDocFiles, localeFlags, localeLabels, mintlifyLocaleToDir, rtlLocales } from "./config.mjs";
@@ -13,8 +12,8 @@ import { createRenderCache } from "./render-cache.mjs";
 import { editSourceUrlForPage, frontmatterSourcePath, readSourceMetadata } from "./edit-source.mjs";
 import { elementsFixture } from "./elements-fixture.mjs";
 import { parseFrontmatter } from "../../.openclaw-sync/lib/docs-markdown.mjs";
-import { renderPageOgSvg } from "./og-card-template.mjs";
-import { createOgCache } from "./og-cache.mjs";
+import { renderPageOgCards } from "./og-cards.mjs";
+import { activeTabTitle, groupForPage, flattenNavEntries, flattenNav } from "./navigation.mjs";
 import { resolveRedirects } from "../../.openclaw-sync/lib/docs-redirects.mjs";
 
 const root = process.cwd();
@@ -35,7 +34,6 @@ const canonicalOrigin = (process.env.DOCS_SITE_CANONICAL_ORIGIN
 const feedbackIssueRepository = normalizeRepository(process.env.DOCS_FEEDBACK_ISSUE_REPO ?? "openclaw/openclaw");
 const llmsFullAvailable = process.env.DOCS_SITE_LLMS_FULL_AVAILABLE === "1";
 const ogImagePath = "/og-card.png";
-const renderedPageOgCards = new Set();
 const chatApiUrl = process.env.DOCS_SITE_CHAT_API_URL ?? "/ask-molty/api/chat";
 const shellCss = siteCss();
 const shellJs = siteJs();
@@ -101,7 +99,10 @@ fs.rmSync(redirectMetadataPath, { force: true });
 fs.rmSync(outDir, { recursive: true, force: true, maxRetries: 5, retryDelay: 100 });
 fs.mkdirSync(outDir, { recursive: true });
 copyPublicFiles();
-if (!previewMode) await renderPageOgCards();
+const renderedPageOgCards = previewMode ? new Set() : await renderPageOgCards({
+  pages, enNav: navByLocale.get("en") ?? [], outDir,
+  cacheDir: path.join(root, ".cache", "docs-og"), siteName: config.name,
+});
 for (const page of pages) writePage(page);
 if (renderCache) {
   renderCache.prune();
@@ -819,88 +820,6 @@ function componentLabel(name, attrs) {
   return label ? `\n${label}\n` : `\n${name}\n`;
 }
 
-async function renderPageOgCards() {
-  const enNav = navByLocale.get("en") ?? [];
-  const navSlugs = collectNavSlugs(enNav);
-  const ogDir = path.join(outDir, "og");
-  const targets = pages.filter((page) =>
-    page.locale === "en" && page.slug !== "index" && navSlugs.has(page.slug)
-  );
-  const start = Date.now();
-  const cache = process.env.DOCS_SITE_RENDER_CACHE !== "0"
-    ? createOgCache(path.join(root, ".cache", "docs-og"), renderOgPng)
-    : null;
-  const concurrency = Math.max(2, Math.min(8, Number(process.env.DOCS_SITE_OG_CONCURRENCY) || 6));
-  let cursor = 0;
-  let count = 0;
-  const failures = [];
-  await Promise.all(Array.from({ length: concurrency }, async () => {
-    while (cursor < targets.length) {
-      const page = targets[cursor++];
-      const kicker = groupForPage(enNav, page.slug) ?? activeTabTitle(enNav, page.slug) ?? config.name;
-      const svg = renderPageOgSvg({ title: page.title, kicker, summary: page.summary });
-      const outFile = path.join(ogDir, `${page.slug}.png`);
-      fs.mkdirSync(path.dirname(outFile), { recursive: true });
-      try {
-        fs.writeFileSync(outFile, await (cache ? cache.render(page.slug, svg) : renderOgPng(svg)));
-        renderedPageOgCards.add(page.slug);
-        count++;
-      } catch (err) {
-        failures.push(`${page.slug}: ${err.message}`);
-      }
-    }
-  }));
-  if (failures.length) {
-    const details = failures.slice(0, 5).join("; ");
-    throw new Error(`failed to render ${failures.length}/${targets.length} per-page og cards: ${details}`);
-  }
-  if (cache) {
-    cache.prune();
-    console.log(`og cache: ${cache.stats.hits} reused, ${cache.stats.misses} rendered`);
-  } else {
-    console.log(`og cache: disabled, ${count} rendered`);
-  }
-  console.log(`prepared ${count}/${targets.length} per-page og cards in ${Date.now() - start}ms`);
-}
-
-function renderOgPng(svg) {
-  return new Promise((resolve, reject) => {
-    const worker = new Worker(new URL("./og-render-worker.mjs", import.meta.url), {
-      workerData: { svg },
-    });
-    let settled = false;
-    worker.on("message", (message) => {
-      if (settled) return;
-      settled = true;
-      if (message?.error) reject(new Error(message.error));
-      else resolve(Buffer.from(message.png));
-    });
-    worker.on("error", (err) => {
-      if (settled) return;
-      settled = true;
-      reject(err);
-    });
-    worker.on("exit", (code) => {
-      if (settled || code === 0) return;
-      settled = true;
-      reject(new Error(`og render worker exit ${code}`));
-    });
-  });
-}
-
-function collectNavSlugs(nav) {
-  const slugs = new Set();
-  for (const tab of nav) {
-    for (const group of tab.groups ?? []) {
-      for (const entry of group.pages ?? []) {
-        if (entry.group) for (const sub of entry.pages ?? []) slugs.add(sub.slug);
-        else if (entry.slug) slugs.add(entry.slug);
-      }
-    }
-  }
-  return slugs;
-}
-
 function writeStaticAssets() {
   const assetsDir = path.join(outDir, "assets");
   fs.mkdirSync(assetsDir, { recursive: true });
@@ -953,28 +872,6 @@ function walkPublicFiles(dir) {
 function copyDir(source, dest, options = {}) {
   if (!fs.existsSync(source)) return;
   fs.cpSync(source, dest, { recursive: true, filter: options.filter });
-}
-
-function activeTabTitle(nav, slug) {
-  return nav.find((tab) => flattenNav([tab]).some((page) => page.slug === slug))?.title ?? nav[0]?.title ?? "";
-}
-
-function groupForPage(nav, slug) {
-  for (const tab of nav) {
-    for (const group of tab.groups) {
-      if (flattenNavEntries(group.pages).some((page) => page.slug === slug)) {
-        return group.title;
-      }
-    }
-  }
-}
-
-function flattenNavEntries(entries) {
-  return entries.flatMap((entry) => entry.group ? flattenNavEntries(entry.pages) : [entry]);
-}
-
-function flattenNav(nav) {
-  return nav.flatMap((tab) => tab.groups.flatMap((group) => flattenNavEntries(group.pages)));
 }
 
 function firstPage(tab) {
